+++ flutter_osm_interface/lib/flutter_osm_interface.dart
@@ -1,0 +1,14 @@
+library flutter_osm_interface;
+
+
+
+export 'src/channel/osm_method_channel.dart' show MethodChannelOSM,config;
+export 'src/osm_interface.dart';
+export 'src/map_controller/base_map_controller.dart';
+export 'src/types/types.dart';
+export 'src/common/utilities.dart';
+export 'src/common/osm_event.dart';
+export 'src/common/geo_point_exception.dart';
+export 'src/common/road_exception.dart';
+export 'src/osm_controller/osm_controller.dart';
+export 'src/map_controller/i_base_map_controller.dart';