+++ flutter_osm_interface/lib/src/common/road_exception.dart
@@ -1,0 +1,11 @@
+class RoadException implements Exception {
+  final String? msg;
+
+  RoadException({
+    this.msg,
+  });
+
+  String? errorMessage() {
+    return msg;
+  }
+}