+++ flutter_osm_interface/lib/src/map_controller/i_base_map_controller.dart
@@ -1,26 +1,8 @@
-library osm_flutter;
-
-import 'dart:async';
-
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/rendering.dart';
-import 'package:location/location.dart';
+import 'package:flutter_osm_interface/flutter_osm_interface.dart';
 
-import '../../flutter_osm_plugin.dart';
-import '../controller/osm_controller.dart';
-import '../types/geo_point.dart';
 import '../types/types.dart';
-import '../widgets/copyright_osm_widget.dart';
-import 'osm_controller.dart';
-
-part '../osm_flutter.dart';
-
-part '../widgets/custom_picker_location.dart';
-
-part 'map_controller.dart';
-
-part 'picker_map_controller.dart';
 
 ///  [BaseMapController] : base controller for osm flutter
 ///
@@ -28,10 +10,7 @@ part 'picker_map_controller.dart';
 /// [initMapWithUserPosition] : (bool) if is true, map will show your current location
 ///
 /// [initPosition] : (GeoPoint) if it isn't null, the map will be pointed at this position
-///
-/// [boundingBox] : (BoundingBox) if it isn't null, the map will be pointed at this position
-abstract class BaseMapController {
-  late OSMController _osmController;
+abstract class IBaseMapController {
   final bool initMapWithUserPosition;
   final GeoPoint? initPosition;
   final BoundingBox? areaLimit;
@@ -48,22 +27,18 @@ abstract class BaseMapController {
 
   ValueListenable<bool> get listenerMapIsReady => _listenerMapIsReady;
 
-  BaseMapController({
+  IBaseMapController({
     this.initMapWithUserPosition = true,
     this.initPosition,
     this.areaLimit = const BoundingBox.world(),
   }) : assert(initMapWithUserPosition || initPosition != null);
 
-  void _init(
-    OSMController osmController,
-  ) {
-    this._osmController = osmController;
-    Future.delayed(Duration(milliseconds: 1250), () async {
-      await this._osmController.initMap(
-          initPosition: initPosition,
-          initWithUserPosition: initMapWithUserPosition,
-          box: areaLimit);
-    });
+  void init();
+
+  void dispose() {
+    _listenerMapLongTapping.dispose();
+    _listenerMapSingleTapping.dispose();
+    _listenerMapIsReady.dispose();
   }
 }