+++ flutter_osm_interface/lib/src/osm_interface.dart
@@ -1,32 +1,25 @@
 import 'package:flutter/material.dart';
-import 'package:flutter/rendering.dart';
-import 'package:flutter_osm_plugin/flutter_osm_plugin.dart';
-import 'package:flutter_osm_plugin/src/channel/osm_method_channel.dart';
-import 'package:location/location.dart';
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
 
+import 'channel/osm_method_channel.dart';
+import 'common/osm_event.dart';
+import 'types/types.dart';
+
 abstract class OSMPlatform extends PlatformInterface {
-  OSMPlatform() : super(token: _token);
-  late Location locationService;
+  OSMPlatform() : super(token: token);
 
-  static final Object _token = Object();
-
-  static OSMPlatform _instance = MethodChannelOSM();
+  static late OSMPlatform _instance = MethodChannelOSM();
 
   static OSMPlatform get instance => _instance;
 
   /// Platform-specific plugins should set this with their own platform-specific
   /// class that extends [OSMPlatformInterface] when they register themselves.
   static set instance(OSMPlatform instance) {
-    PlatformInterface.verifyToken(instance, _token);
+    PlatformInterface.verifyToken(instance, token);
     _instance = instance;
   }
 
-  Future<void> init(
-    int idOSM,
-  );
-
-  void close();
+  static final Object token = Object();
 
   Stream<MapInitialization> onMapIsReady(int idMap);
 
@@ -38,6 +31,22 @@ abstract class OSMPlatform extends PlatformInterface {
 
   Stream<UserLocationEvent> onUserPositionListener(int idMap);
 
+  Future<void> init(
+    int idOSM,
+  );
+
+  void close();
+
+// Future<void> configureZoomMap(
+//   int idOSM,
+//   double initZoom,
+//   int minZoomLevel,
+//   int maxZoomLevel,
+//   double stepZoom,
+// );
+}
+
+abstract class MobileOSMPlatform extends OSMPlatform {
   Future<void> initMap(
     int idOSM,
     GeoPoint point,